Polishing merge (#19)
* Update admincog.py

* Update main.py

* Update admincog.py

* Update utilscog.py

* merge (#15)

* Update README.md

* Update README.md

* Update __init__.py

* Update __init__.py

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* merge (#16)

* Update README.md

* Update README.md

* Update __init__.py

* Update __init__.py

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update requestcog.py

* added werewolves channel

* Update spotcog.py

* Update spotcog.py

* Update requirements.txt

* embed color from mebmer pfp

* Update leaderboardcog.py

* Update utils.py

* Update leaderboardcog.py

* Update spotcog.py

* Update leaderboardcog.py

* no emotes on #nemeton

* fixed permissions security issue

* added changing spots command

* fixed error

* added command for changing monster type

* Update utilscog.py

* Update utilscog.py

* Update admincog.py

* Update utilscog.py

* Update utilscog.py

* Update admincog.py

* Update admincog.py

* Update utilscog.py

* small fixes

* Update main.py

* Update requestcog.py

* Update utilscog.py

* Update main.py

* Update main.py

* Update cogbase.py

* Update cogbase.py

* Create spotstatscog.py

* Update utilscog.py

* Update spotstatscog.py

* Update main.py

* Update spotstatscog.py

* Update spotstatscog.py

* Update spotstatscog.py

* Update requestcog.py

* Update spotstatscog.py

* Update leaderboardcog.py

* Update spotstatscog.py

* Update requestcog.py

* Update admincog.py

* Update main.py

* Update leaderboardcog.py

* Update spotstatscog.py

* Update spotstatscog.py

* Update spotstatscog.py

* Update spotstatscog.py

* Update cogbase.py

* Update admincog.py

* improved logging

* Update utilscog.py

* Update commandscog.py

* Update utilscog.py

* hotfixes

* Update admincog.py

* Update spotcog.py

* Update spotcog.py

* Update spotcog.py

* Update utilscog.py

* Update utilscog.py

* updated commands desc

* code refactoring

* time display mode changed

* Update requestcog.py

* added legend multiplier

* new stats command formatting

* small code fixes

* purge error handle

* small naming change

* Create coordscog.py

* updated message

* Update coordscog.py

+++ cogs/spotcog.py
@@ -5,7 +5,7 @@ Current commands:
 /remove_spot
 
 """
-
+import discord
 from discord.ext import commands
 from discord.utils import get
 import cogs.cogbase as cogbase
@@ -20,6 +20,7 @@ cords_beginning = ["-", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
 class SpotCog(cogbase.BaseCog):
     def __init__(self, base):
         super().__init__(base)
+        self.peepo_ban_emote = ":peepoban:872502800146382898"
 
     # Ping monster role
     @commands.Cog.listener()
@@ -36,8 +37,8 @@ class SpotCog(cogbase.BaseCog):
     @staticmethod
     async def handle_spotted_common(ctx):
         if ctx.content[0] in cords_beginning:
-            await DatabaseCog.db_count_spot(ctx.author.id, monster_type_dict[4])
-            await DatabaseCog.db_save_coords(ctx.content, monster_type_dict[4])
+            await DatabaseCog.db_count_spot(ctx.author.id, "common")
+            await DatabaseCog.db_save_coords(ctx.content, "common")
         else:
             await ctx.delete()
 
@@ -46,22 +47,37 @@ class SpotCog(cogbase.BaseCog):
         if ctx.content.startswith(prefix):
             spotted_monster = self.get_monster(ctx, ctx.content.replace(prefix, ""))
             if spotted_monster:
+                monster_type_str = monster_type_dict[spotted_monster["type"]]
+                if await self.wrong_channel(ctx, spotted_monster, monster_type_str):
+                    return
                 role = get(ctx.guild.roles, name=spotted_monster["name"])
                 await ctx.delete()
                 await ctx.channel.send(f"{role.mention}")
                 await DatabaseCog.db_count_spot(ctx.author.id,
-                                                monster_type_dict[spotted_monster["type"]])
+                                                monster_type_str)
                 logs_ch = self.bot.get_channel(self.bot.ch_logs)
                 await logs_ch.send(f"[PingLog] {ctx.author} ({ctx.author.id}) "
-                                   f"requested ping for **{spotted_monster['name']}** role")
+                                   f"requested ping for **{spotted_monster['name']}**")
             else:
                 await ctx.delete()
                 await ctx.channel.send(
                     f"{ctx.author.mention} monster not found - are you sure that the name is correct?", delete_after=5)
         elif len(ctx.content) > 0 and ctx.content[0] in cords_beginning:
             await DatabaseCog.db_save_coords(ctx.content, ctx.channel.name)
-        elif ctx.channel.id != self.bot.ch_werewolf and ctx.channel.id != self.bot.ch_nemeton:
-            await ctx.add_reaction("a:peepoban:872502800146382898")
+        elif ctx.channel.id == self.bot.ch_legendary_spot or ctx.channel.id == self.bot.ch_rare_spot:
+            await ctx.add_reaction(f"a{self.peepo_ban_emote}")
+
+    async def wrong_channel(self, ctx, spotted_monster, monster_type_str):
+        if ctx.channel.id in [self.bot.ch_legendary_spot, self.bot.ch_rare_spot]:
+            if ctx.channel.name != monster_type_str:
+                channel = discord.utils.get(ctx.guild.channels, name=monster_type_str)
+                correct_channel = channel.id
+                await ctx.delete()
+                await ctx.channel.send(
+                    f"{ctx.author.mention} you posted {spotted_monster['name']} on wrong channel! "
+                    f"Use <#{correct_channel}> instead! <{self.peepo_ban_emote}>", delete_after=8)
+                return True
+            return False
 
 
 def setup(bot: commands.Bot):
